Move AnnounceButton to own file
